feat: remove dependicy on humanize-plus
